Split source formatting into a separate file
